+++ tests/test_authentication.py
@@ -1,40 +1,43 @@
 from __future__ import annotations
 
+import time
 import typing as t
+from unittest import mock
 
 import globus_sdk
 import pytest
-from globus_sdk._testing import load_response
+from globus_sdk._testing import RegisteredResponse, load_response
 
-from globus_action_provider_tools.authentication import AuthState, identity_principal
+from globus_action_provider_tools.authentication import (
+    AuthState,
+    InvalidTokenScopesError,
+    identity_principal,
+)
 
 
-def get_auth_state_instance(
-    expected_scopes: t.Iterable[str],
-    expected_audience: str,
-) -> AuthState:
-    return AuthState(
-        auth_client=globus_sdk.ConfidentialAppAuthClient("bogus", "bogus"),
-        bearer_token="bogus",
-        expected_scopes=expected_scopes,
-        expected_audience=expected_audience,
+def get_auth_state_instance(expected_scopes: t.Iterable[str]) -> AuthState:
+    client = globus_sdk.ConfidentialAppAuthClient(
+        "bogus", "bogus", transport_params={"max_retries": 0}
     )
+    return AuthState(
+        auth_client=client,
+        bearer_token="bogus",
+        expected_scopes=frozenset(expected_scopes),
+    )
+
+
+@pytest.fixture(autouse=True)
+def _clear_auth_state_cache():
+    AuthState.dependent_tokens_cache.clear()
+    AuthState.group_membership_cache.clear()
+    AuthState.introspect_cache.clear()
 
 
 @pytest.fixture
 def auth_state(mocked_responses) -> AuthState:
-    """Create an AuthState instance.
-
-    AuthState compares its `expected_scopes` and `expected_audience` values
-    against the values present in an API request and will fail if they don't match.
-    Unfortunately, this currently means that these values are duplicated
-    in the API fixture .yaml files and here.
-    """
-
-    AuthState.dependent_tokens_cache.clear()
-    AuthState.group_membership_cache.clear()
-    AuthState.introspect_cache.clear()
-    return get_auth_state_instance(["expected-scope"], "expected-audience")
+    """Create an AuthState instance."""
+    # note that expected-scope MUST match the fixture data
+    return get_auth_state_instance(["expected-scope"])
 
 
 def test_get_identities(auth_state, freeze_time):
@@ -84,10 +87,7 @@ def test_caching_groups(auth_state, freeze_time, mocked_responses):
 def test_auth_state_caching_across_instances(auth_state, freeze_time, mocked_responses):
     response = freeze_time(load_response("token-introspect", case="success"))
 
-    duplicate_auth_state = get_auth_state_instance(
-        auth_state.expected_scopes,
-        auth_state.expected_audience,
-    )
+    duplicate_auth_state = get_auth_state_instance(auth_state.expected_scopes)
     assert duplicate_auth_state is not auth_state
 
     assert len(auth_state.identities) == len(response.metadata["identities"])
@@ -102,3 +102,77 @@ def test_invalid_grant_exception(auth_state):
     load_response("token-introspect", case="success")
     load_response("token", case="invalid-grant")
     assert auth_state.get_authorizer_for_scope("doesn't matter") is None
+
+
+def test_dependent_token_callout_500_fails_dependent_authorization(auth_state):
+    """
+    On a 5xx response, getting an authorizer fails.
+
+    FIXME: currently this simply emits 'None' -- in the future the error should propagate
+    """
+    RegisteredResponse(
+        service="auth", path="/v2/oauth2/token", method="POST", status=500
+    ).add()
+    assert (
+        auth_state.get_authorizer_for_scope(
+            "urn:globus:auth:scope:groups.api.globus.org:view_my_groups_and_memberships"
+        )
+        is None
+    )
+
+
+def test_dependent_token_callout_success_fixes_bad_cache(auth_state):
+    """
+    Populate the cache "incorrectly" and then "fix it" by asking for an authorizer
+    and expecting the dependent token logic to appropriately redrive.
+    """
+    # the mock by_scopes value is a dict -- similar enough since it just needs to be
+    # a mapping type -- which we populate for "foo"
+    mock_response = mock.Mock()
+    mock_response.by_scopes = {
+        "foo_scope": {
+            "expires_at_seconds": time.time() + 100,
+            "access_token": "foo_AT",
+            "refresh_token": "foo_RT",
+        }
+    }
+    auth_state.dependent_tokens_cache[auth_state._dependent_token_cache_key] = (
+        mock_response
+    )
+
+    # register a response for a different resource server -- 'bar'
+    RegisteredResponse(
+        service="auth",
+        path="/v2/oauth2/token",
+        method="POST",
+        json=[
+            {
+                "resource_server": "bar",
+                "scope": "bar_scope",
+                "expires_at_seconds": time.time() + 100,
+                "access_token": "bar_AT",
+                "refresh_token": "bar_RT",
+            }
+        ],
+    ).add()
+    # now get the 'bar_scope' authorizer
+    authorizer = auth_state.get_authorizer_for_scope("bar_scope")
+
+    # it should be a refresh token authorizer and the cache should be updated
+    assert isinstance(authorizer, globus_sdk.RefreshTokenAuthorizer)
+    cache_value = auth_state.dependent_tokens_cache[
+        auth_state._dependent_token_cache_key
+    ]
+    assert isinstance(cache_value, globus_sdk.OAuthDependentTokenResponse)
+    assert "foo_scope" not in cache_value.by_scopes
+    assert "bar_scope" in cache_value.by_scopes
+
+
+def test_invalid_scopes_error():
+    auth_state = get_auth_state_instance(["bad-scope"])
+    load_response("token-introspect", case="success")
+    with pytest.raises(InvalidTokenScopesError) as excinfo:
+        auth_state.introspect_token()
+
+    assert excinfo.value.expected_scopes == {"bad-scope"}
+    assert excinfo.value.actual_scopes == {"expected-scope"}